% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_country.R
\name{country_module_sidebar}
\alias{country_module_sidebar}
\title{Country module Sidebar}
\usage{
country_module_sidebar(id)
}
\arguments{
\item{id}{the ID of the module}
}
\description{
Country module Sidebar
}
\keyword{internal}
