% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_country.R
\name{country_module_ui}
\alias{country_module_ui}
\title{Country module UI}
\usage{
country_module_ui(id)
}
\arguments{
\item{id}{the ID of the module}
}
\description{
Country module UI
}
\keyword{internal}
