% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_istraxfunctions.R
\name{get_prepdata_path}
\alias{get_prepdata_path}
\title{Get the prepdata directory path}
\usage{
get_prepdata_path(iseapp_local_path)
}
\arguments{
\item{iseapp_local_path}{The base iseapp path}
}
\value{
Path to prepdata directory or NULL
}
\description{
Get the prepdata directory path
}
