% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_globe.R
\name{globe_module_sidebar}
\alias{globe_module_sidebar}
\title{Globe module Sidebar}
\usage{
globe_module_sidebar(id)
}
\arguments{
\item{id}{the ID of the module}
}
\description{
Globe module Sidebar
}
\keyword{internal}
