% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_home_module.R
\name{home_module_server}
\alias{home_module_server}
\title{Home module Server}
\usage{
home_module_server(id)
}
\arguments{
\item{id}{the ID of the module}
}
\description{
Home module Server
}
\keyword{internal}
