% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innovationStrategyExplorer-package.R
\docType{package}
\name{innovationStrategyExplorer-package}
\alias{innovationStrategyExplorer}
\alias{innovationStrategyExplorer-package}
\title{innovationStrategyExplorer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
