% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data.R
\name{load_big_datasets}
\alias{load_big_datasets}
\title{Load big datasets}
\usage{
load_big_datasets()
}
\value{
A list containing techmap, countrymap, and regionmap
}
\description{
Load big datasets
}
\keyword{internal}
