% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_istraxfunctions.R
\name{load_precomputed_by_region}
\alias{load_precomputed_by_region}
\title{Try to load pre-computed data for "by region" aggregation}
\usage{
load_precomputed_by_region(prepdata_path, toflow, tech_category_name)
}
\arguments{
\item{prepdata_path}{Path to prepdata directory}

\item{toflow}{The toflow variable name}

\item{tech_category_name}{Name of the technology category}
}
\value{
Data frame or NULL if not found
}
\description{
Try to load pre-computed data for "by region" aggregation
}
