% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_istraxfunctions.R
\name{load_precomputed_by_tech}
\alias{load_precomputed_by_tech}
\title{Try to load pre-computed data for "by technology" aggregation}
\usage{
load_precomputed_by_tech(toflow, country_group_name)
}
\arguments{
\item{toflow}{The toflow variable name}

\item{country_group_name}{Name of the country group}
}
\value{
Data frame or NULL if not found
}
\description{
Try to load pre-computed data for "by technology" aggregation
}
