% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_istraxfunctions.R
\name{match_country_group}
\alias{match_country_group}
\title{Map selected countries to a country group name for pre-computed lookup}
\usage{
match_country_group(selected_countries, group_definitions)
}
\arguments{
\item{selected_countries}{Vector of selected country codes}

\item{group_definitions}{Named list of country group definitions}
}
\value{
Group name if exact match found, NULL otherwise
}
\description{
Map selected countries to a country group name for pre-computed lookup
}
