% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_other_module.R
\name{other_module_server}
\alias{other_module_server}
\title{Other module Server}
\usage{
other_module_server(id)
}
\arguments{
\item{id}{the ID of the module}
}
\description{
Other module Server
}
\keyword{internal}
