% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{plotly_title_style}
\alias{plotly_title_style}
\title{Helper for plotly title styling}
\usage{
plotly_title_style(title_size = 24, title_color = "#4b4b4b")
}
\arguments{
\item{title_size}{Size for plot title (default 18)}

\item{title_color}{Color for title text (default "#4b4b4b")}
}
\value{
A list with plotly title parameters
}
\description{
Returns a list of plotly title layout parameters
}
\keyword{internal}
