% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_region.R
\name{region_module_sidebar}
\alias{region_module_sidebar}
\title{Region module Sidebar}
\usage{
region_module_sidebar(id)
}
\arguments{
\item{id}{the ID of the module}
}
\description{
Region module Sidebar
}
\keyword{internal}
