% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_url_restore.R
\name{restore_module_inputs}
\alias{restore_module_inputs}
\title{Generic URL parameter restoration for modules}
\usage{
restore_module_inputs(params, module_prefix, tab_name, input_configs, session)
}
\arguments{
\item{params}{List of URL parameters}

\item{module_prefix}{Module namespace prefix (e.g., "country-", "region-")}

\item{tab_name}{Expected tab name (e.g., "Country Explorer")}

\item{input_configs}{Named list mapping input IDs to their types}

\item{session}{Module session object}
}
\description{
Generic URL parameter restoration for modules
}
\keyword{internal}
