% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_istraxfunctions.R
\name{safe_filename}
\alias{safe_filename}
\title{Generate safe filename for pre-computed data}
\usage{
safe_filename(...)
}
\arguments{
\item{...}{Parts to combine into filename}
}
\value{
Safe filename string
}
\description{
Generate safe filename for pre-computed data
}
