% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny app Server}
\usage{
server(input, output, session)
}
\arguments{
\item{input, output}{Input and output list objects
containing said registered inputs and outputs.}

\item{session}{Shiny session.}
}
\description{
Core server function.
}
\keyword{internal}
