% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_country.R
\name{country_module_server}
\alias{country_module_server}
\title{Country module Server}
\usage{
country_module_server(id, parent_session)
}
\arguments{
\item{id}{the ID of the module}
}
\description{
Country module Server
}
\keyword{internal}
