% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_globe.R
\name{globe_module_server}
\alias{globe_module_server}
\title{Globe module Server}
\usage{
globe_module_server(id)
}
\arguments{
\item{id}{the ID of the module}
}
\description{
Globe module Server
}
\keyword{internal}
