% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_region.R
\name{region_module_server}
\alias{region_module_server}
\title{Region module Server}
\usage{
region_module_server(id)
}
\arguments{
\item{id}{the ID of the module}
}
\description{
Region module Server
}
\keyword{internal}
